refactor: Add battle event queue
We need a way to process animations and actions that take place during a
battle. Rather than have a ton of separate members, a queue seemed like
the most generic approach. This change modifies Battle to run updates
and renders on items in the queue synchronously. A common interface for
this is shared by other members - the battle animation and dialogue.
These members have been added to the queue.