Decouple BattleMenu offset from Actor's position
There was a rendering offset placed on the BattleMenu's rendering which
would cause the menu to follow the Actor when it moved in to attack a
foe. We would rather that the menu stay still in this scenario which
required us to define a mutable position on the base Menu class. This
was the most straight-forward way. Since there are many pre-existing
menus we will default the position to [0, 0] as to not affect other
menus. This change also means that we will have to update the
BattleMenu's position in the future when there are more than one member
on the team because swapping between members will need the BattleMenu to
relocate underneath their sprite.