refactor: Normalize string ref to battle menu items
The battle menu contains a variety of items from weapons, to abilities,
to one-off, default tactics such as defending and running away. This
change normalizes the programatic way of referring to them by ref
(previously called type). This change also structures the ability
manifest to align with an expected interface.