feat: Allow multiple players to participate in battle
Previously the other players would render in the battle but not be able
to take their turn. This change allows all players to take their turn.
However, the current design of the menu system is blocking the menus
from properly rendering the weapons/abilities of other players besides
the main one.