fix: Fix unequipped item bug
When loading members from the state on game init, the state was resaving
the players' states before all the init code had run and clear out their
equipped items. This change targets how the overall team gets saved to
the state by cherry-picking the state of a given player when added to
the team and only updates the state for that player.